# [Core] Optimize `OAHashMap`