--- a/core/templates/oa_hash_map.h
+++ b/core/templates/oa_hash_map.h
@@ -35,6 +35,12 @@
 #include "core/os/memory.h"
 #include "core/templates/hashfuncs.h"
 
+template <typename K, typename V>
+struct OAHashMapElement {
+	K key;
+	V value;
+};
+
 /**
  * A HashMap implementation that uses open addressing with Robin Hood hashing.
  * Robin Hood hashing swaps out entries that have a smaller probing distance
@@ -55,10 +61,10 @@ template <typename TKey, typename TValue,
 		typename Comparator = HashMapComparatorDefault<TKey>>
 class OAHashMap {
 private:
-	TValue *values = nullptr;
-	TKey *keys = nullptr;
+	OAHashMapElement<TKey, TValue> *elements = nullptr;
 	uint32_t *hashes = nullptr;
 
+	// Due to optimization, this is `capacity - 1`. Use + 1 to get normal capacity.
 	uint32_t capacity = 0;
 
 	uint32_t num_elements = 0;
@@ -68,116 +74,174 @@ private:
 	_FORCE_INLINE_ uint32_t _hash(const TKey &p_key) const {
 		uint32_t hash = Hasher::hash(p_key);
 
-		if (hash == EMPTY_HASH) {
+		if (unlikely(hash == EMPTY_HASH)) {
 			hash = EMPTY_HASH + 1;
 		}
 
 		return hash;
 	}
 
-	_FORCE_INLINE_ uint32_t _get_probe_length(uint32_t p_pos, uint32_t p_hash) const {
-		uint32_t original_pos = p_hash % capacity;
-		return (p_pos - original_pos + capacity) % capacity;
+	static _FORCE_INLINE_ uint32_t _get_resize_count(uint32_t p_capacity) {
+		return p_capacity ^ (p_capacity + 1) >> 2; // = get_capacity() * 0.75 - 1; Works only if p_capacity = 2^n - 1.
+	}
+
+	static _FORCE_INLINE_ uint32_t _get_probe_length(uint32_t p_pos, uint32_t p_hash, uint32_t p_local_capacity) {
+		const uint32_t original_pos = p_hash & p_local_capacity;
+		return (p_pos - original_pos + p_local_capacity + 1) & p_local_capacity;
 	}
 
 	_FORCE_INLINE_ void _construct(uint32_t p_pos, uint32_t p_hash, const TKey &p_key, const TValue &p_value) {
-		memnew_placement(&keys[p_pos], TKey(p_key));
-		memnew_placement(&values[p_pos], TValue(p_value));
+		if constexpr (!std::is_trivially_constructible_v<TKey>) {
+			memnew_placement(&elements[p_pos].key, TKey(p_key));
+		} else {
+			TKey key = p_key;
+			elements[p_pos].key = key;
+		}
+		if constexpr (!std::is_trivially_constructible_v<TValue>) {
+			memnew_placement(&elements[p_pos].value, TValue(p_value));
+		} else {
+			TValue value = p_value;
+			elements[p_pos].value = value;
+		}
+
 		hashes[p_pos] = p_hash;
 
 		num_elements++;
 	}
 
-	bool _lookup_pos(const TKey &p_key, uint32_t &r_pos) const {
-		uint32_t hash = _hash(p_key);
-		uint32_t pos = hash % capacity;
-		uint32_t distance = 0;
+	_FORCE_INLINE_ bool _lookup_pos(const TKey &p_key, uint32_t &r_pos) const {
+		return _lookup_pos_with_hash(p_key, r_pos, _hash(p_key));
+	}
 
+	_FORCE_INLINE_ bool _lookup_pos_with_hash(const TKey &p_key, uint32_t &r_pos, uint32_t p_hash) const {
+		uint32_t pos = p_hash & capacity;
+
+		if (hashes[pos] == p_hash && Comparator::compare(elements[pos].key, p_key)) {
+			r_pos = pos;
+			return true;
+		}
+
+		if (hashes[pos] == EMPTY_HASH) {
+			return false;
+		}
+
+		// A collision occurred.
+		pos = (pos + 1) & capacity;
+		uint32_t distance = 1;
 		while (true) {
-			if (hashes[pos] == EMPTY_HASH) {
-				return false;
-			}
-
-			if (distance > _get_probe_length(pos, hashes[pos])) {
-				return false;
-			}
-
-			if (hashes[pos] == hash && Comparator::compare(keys[pos], p_key)) {
+			if (hashes[pos] == p_hash && Comparator::compare(elements[pos].key, p_key)) {
 				r_pos = pos;
 				return true;
 			}
 
-			pos = (pos + 1) % capacity;
+			if (hashes[pos] == EMPTY_HASH) {
+				return false;
+			}
+
+			if (distance > _get_probe_length(pos, hashes[pos], capacity)) {
+				return false;
+			}
+
+			pos = (pos + 1) & capacity;
 			distance++;
 		}
 	}
 
-	void _insert_with_hash(uint32_t p_hash, const TKey &p_key, const TValue &p_value) {
+	_FORCE_INLINE_ void _insert_with_swaps(uint32_t p_hash, uint32_t p_pos, uint32_t p_distance, const TKey &p_key, const TValue &p_value) {
 		uint32_t hash = p_hash;
-		uint32_t distance = 0;
-		uint32_t pos = hash % capacity;
-
+		uint32_t distance = p_distance;
+		uint32_t pos = p_pos;
 		TKey key = p_key;
 		TValue value = p_value;
-
 		while (true) {
-			if (hashes[pos] == EMPTY_HASH) {
-				_construct(pos, hash, key, value);
-
-				return;
-			}
-
-			// not an empty slot, let's check the probing length of the existing one
-			uint32_t existing_probe_len = _get_probe_length(pos, hashes[pos]);
+			uint32_t existing_probe_len = _get_probe_length(pos, hashes[pos], capacity);
 			if (existing_probe_len < distance) {
 				SWAP(hash, hashes[pos]);
-				SWAP(key, keys[pos]);
-				SWAP(value, values[pos]);
+				SWAP(key, elements[pos].key);
+				SWAP(value, elements[pos].value);
 				distance = existing_probe_len;
 			}
 
-			pos = (pos + 1) % capacity;
+			pos = (pos + 1) & capacity;
+			distance++;
+
+			if (hashes[pos] == EMPTY_HASH) {
+				_construct(pos, hash, key, value);
+				return;
+			}
+		}
+	}
+
+	_FORCE_INLINE_ void _insert_with_hash(uint32_t p_hash, const TKey &p_key, const TValue &p_value) {
+		uint32_t pos = p_hash & capacity;
+
+		if (hashes[pos] == EMPTY_HASH) {
+			_construct(pos, p_hash, p_key, p_value);
+			return;
+		}
+
+		// A collision occurred.
+		pos = (pos + 1) & capacity;
+		uint32_t distance = 1;
+		while (true) {
+			if (hashes[pos] == EMPTY_HASH) {
+				_construct(pos, p_hash, p_key, p_value);
+				return;
+			}
+
+			// Not an empty slot, let's check the probing length of the existing one.
+			uint32_t existing_probe_len = _get_probe_length(pos, hashes[pos], capacity);
+			if (existing_probe_len < distance) {
+				// Now we force to make key-value copy.
+				_insert_with_swaps(p_hash, pos, distance, p_key, p_value);
+				return;
+			}
+
+			pos = (pos + 1) & capacity;
 			distance++;
 		}
 	}
 
 	void _resize_and_rehash(uint32_t p_new_capacity) {
-		uint32_t old_capacity = capacity;
+		uint32_t old_real_capacity = capacity + 1;
 
-		// Capacity can't be 0.
-		capacity = MAX(1u, p_new_capacity);
+		// Capacity can't be 0 and must be 2^n - 1.
+		capacity = MAX(4u, p_new_capacity);
+		uint32_t real_capacity = next_power_of_2(capacity);
+		capacity = real_capacity - 1;
 
-		TKey *old_keys = keys;
-		TValue *old_values = values;
+		OAHashMapElement<TKey, TValue> *old_elements = elements;
 		uint32_t *old_hashes = hashes;
 
-		num_elements = 0;
-		keys = static_cast<TKey *>(Memory::alloc_static(sizeof(TKey) * capacity));
-		values = static_cast<TValue *>(Memory::alloc_static(sizeof(TValue) * capacity));
-		hashes = static_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * capacity));
+		hashes = static_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * real_capacity));
+		elements = static_cast<OAHashMapElement<TKey, TValue> *>(Memory::alloc_static(sizeof(OAHashMapElement<TKey, TValue>) * real_capacity));
 
-		for (uint32_t i = 0; i < capacity; i++) {
-			hashes[i] = 0;
-		}
+		memset(hashes, EMPTY_HASH, sizeof(uint32_t) * real_capacity);
 
-		if (old_capacity == 0) {
+		if (old_elements == nullptr) {
 			// Nothing to do.
 			return;
 		}
 
-		for (uint32_t i = 0; i < old_capacity; i++) {
-			if (old_hashes[i] == EMPTY_HASH) {
-				continue;
+		if (num_elements != 0) {
+			num_elements = 0;
+
+			for (uint32_t i = 0; i < old_real_capacity; i++) {
+				if (old_hashes[i] == EMPTY_HASH) {
+					continue;
+				}
+
+				_insert_with_hash(old_hashes[i], old_elements[i].key, old_elements[i].value);
+
+				if constexpr (!std::is_trivially_destructible_v<TKey>) {
+					old_elements[i].key.~TKey();
+				}
+				if constexpr (!std::is_trivially_destructible_v<TValue>) {
+					old_elements[i].value.~TValue();
+				}
 			}
-
-			_insert_with_hash(old_hashes[i], old_keys[i], old_values[i]);
-
-			old_keys[i].~TKey();
-			old_values[i].~TValue();
 		}
-
-		Memory::free_static(old_keys);
-		Memory::free_static(old_values);
+		Memory::free_static(old_elements);
 		Memory::free_static(old_hashes);
 	}
 
@@ -185,30 +249,46 @@ private:
 		_resize_and_rehash(capacity * 2);
 	}
 
+	void _clear_elements() {
+		uint32_t real_capacity = capacity + 1;
+		for (uint32_t i = 0; i < real_capacity; i++) {
+			if (hashes[i] == EMPTY_HASH) {
+				continue;
+			}
+			if constexpr (!std::is_trivially_destructible_v<TValue>) {
+				elements[i].value.~TValue();
+			}
+			if constexpr (!std::is_trivially_destructible_v<TKey>) {
+				elements[i].key.~TKey();
+			}
+		}
+	}
+
+	void _reset() {
+		if (num_elements != 0) {
+			_clear_elements();
+		}
+
+		Memory::free_static(hashes);
+		Memory::free_static(elements);
+	}
+
 public:
-	_FORCE_INLINE_ uint32_t get_capacity() const { return capacity; }
+	_FORCE_INLINE_ uint32_t get_capacity() const { return capacity + 1; }
 	_FORCE_INLINE_ uint32_t get_num_elements() const { return num_elements; }
 
-	bool is_empty() const {
+	_FORCE_INLINE_ bool is_empty() const {
 		return num_elements == 0;
 	}
 
 	void clear() {
-		for (uint32_t i = 0; i < capacity; i++) {
-			if (hashes[i] == EMPTY_HASH) {
-				continue;
-			}
-
-			hashes[i] = EMPTY_HASH;
-			values[i].~TValue();
-			keys[i].~TKey();
-		}
-
+		_clear_elements();
+		memset(hashes, EMPTY_HASH, (capacity + 1) * sizeof(uint32_t));
 		num_elements = 0;
 	}
 
 	void insert(const TKey &p_key, const TValue &p_value) {
-		if (num_elements + 1 > 0.9 * capacity) {
+		if (unlikely(num_elements > _get_resize_count(capacity))) {
 			_resize_and_rehash();
 		}
 
@@ -219,19 +299,23 @@ public:
 
 	void set(const TKey &p_key, const TValue &p_data) {
 		uint32_t pos = 0;
-		bool exists = _lookup_pos(p_key, pos);
+		const uint32_t hash = _hash(p_key);
+		bool exists = _lookup_pos_with_hash(p_key, pos, hash);
 
 		if (exists) {
-			values[pos] = p_data;
+			elements[pos].value = p_data;
 		} else {
-			insert(p_key, p_data);
+			if (unlikely(num_elements > _get_resize_count(capacity))) {
+				_resize_and_rehash();
+			}
+			_insert_with_hash(hash, p_key, p_data);
 		}
 	}
 
 	/**
-	 * returns true if the value was found, false otherwise.
+	 * Returns true if the value was found, false otherwise.
 	 *
-	 * if r_data is not nullptr then the value will be written to the object
+	 * If r_data is not nullptr then the value will be written to the object
 	 * it points to.
 	 */
 	bool lookup(const TKey &p_key, TValue &r_data) const {
@@ -239,7 +323,7 @@ public:
 		bool exists = _lookup_pos(p_key, pos);
 
 		if (exists) {
-			r_data = values[pos];
+			r_data = elements[pos].value;
 			return true;
 		}
 
@@ -251,7 +335,7 @@ public:
 		bool exists = _lookup_pos(p_key, pos);
 
 		if (exists) {
-			return &values[pos];
+			return &elements[pos].value;
 		}
 		return nullptr;
 	}
@@ -261,12 +345,12 @@ public:
 		bool exists = _lookup_pos(p_key, pos);
 
 		if (exists) {
-			return &values[pos];
+			return &elements[pos].value;
 		}
 		return nullptr;
 	}
 
-	_FORCE_INLINE_ bool has(const TKey &p_key) const {
+	bool has(const TKey &p_key) const {
 		uint32_t _pos = 0;
 		return _lookup_pos(p_key, _pos);
 	}
@@ -279,30 +363,34 @@ public:
 			return;
 		}
 
-		uint32_t next_pos = (pos + 1) % capacity;
+		uint32_t next_pos = (pos + 1) & capacity;
 		while (hashes[next_pos] != EMPTY_HASH &&
-				_get_probe_length(next_pos, hashes[next_pos]) != 0) {
+				_get_probe_length(next_pos, hashes[next_pos], capacity) != 0) {
 			SWAP(hashes[next_pos], hashes[pos]);
-			SWAP(keys[next_pos], keys[pos]);
-			SWAP(values[next_pos], values[pos]);
+			SWAP(elements[next_pos].key, elements[pos].key);
+			SWAP(elements[next_pos].value, elements[pos].value);
 			pos = next_pos;
-			next_pos = (pos + 1) % capacity;
+			next_pos = (pos + 1) & capacity;
 		}
 
 		hashes[pos] = EMPTY_HASH;
-		values[pos].~TValue();
-		keys[pos].~TKey();
+		if constexpr (!std::is_trivially_destructible_v<TValue>) {
+			elements[pos].value.~TValue();
+		}
+		if constexpr (!std::is_trivially_destructible_v<TKey>) {
+			elements[pos].key.~TKey();
+		}
 
 		num_elements--;
 	}
 
 	/**
-	 * reserves space for a number of elements, useful to avoid many resizes and rehashes
+	 * Reserves space for a number of elements, useful to avoid many resizes and rehashes
 	 *  if adding a known (possibly large) number of elements at once, must be larger than old
 	 *  capacity.
 	 **/
 	void reserve(uint32_t p_new_capacity) {
-		ERR_FAIL_COND(p_new_capacity < capacity);
+		ERR_FAIL_COND_MSG(p_new_capacity < get_capacity(), "It is impossible to reserve less capacity than is currently available.");
 		_resize_and_rehash(p_new_capacity);
 	}
 
@@ -336,8 +424,8 @@ public:
 		it.pos = p_iter.pos;
 		it.key = nullptr;
 		it.value = nullptr;
-
-		for (uint32_t i = it.pos; i < capacity; i++) {
+		uint32_t real_capacity = capacity + 1;
+		for (uint32_t i = it.pos; i < real_capacity; i++) {
 			it.pos = i + 1;
 
 			if (hashes[i] == EMPTY_HASH) {
@@ -345,8 +433,8 @@ public:
 			}
 
 			it.valid = true;
-			it.key = &keys[i];
-			it.value = &values[i];
+			it.key = &elements[i].key;
+			it.value = &elements[i].value;
 			return it;
 		}
 
@@ -358,43 +446,40 @@ public:
 	}
 
 	void operator=(const OAHashMap &p_other) {
-		if (capacity != 0) {
-			clear();
+		if (&p_other == this) {
+			return;
+		}
+		if (elements != nullptr) {
+			_reset();
 		}
 
-		_resize_and_rehash(p_other.capacity);
+		capacity = p_other.capacity;
+		num_elements = p_other.num_elements;
+		uint32_t real_capacity = capacity + 1;
+		hashes = static_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * real_capacity));
+		elements = static_cast<OAHashMapElement<TKey, TValue> *>(Memory::alloc_static(sizeof(OAHashMapElement<TKey, TValue>) * real_capacity));
 
-		for (Iterator it = p_other.iter(); it.valid; it = p_other.next_iter(it)) {
-			set(*it.key, *it.value);
+		memcpy(hashes, p_other.hashes, sizeof(uint32_t) * real_capacity);
+
+		if constexpr (std::is_trivially_copyable_v<TKey> && std::is_trivially_copyable_v<TValue>) {
+			memcpy(elements, p_other.elements, sizeof(OAHashMapElement<TKey, TValue>) * num_elements);
+		} else {
+			for (uint32_t i = 0; i < real_capacity; i++) {
+				if (hashes[i] == EMPTY_HASH) {
+					continue;
+				}
+				memnew_placement(&elements[i].key, TKey(p_other.elements[i].key));
+				memnew_placement(&elements[i].value, TValue(p_other.elements[i].value));
+			}
 		}
 	}
 
-	OAHashMap(uint32_t p_initial_capacity = 64) {
-		// Capacity can't be 0.
-		capacity = MAX(1u, p_initial_capacity);
-
-		keys = static_cast<TKey *>(Memory::alloc_static(sizeof(TKey) * capacity));
-		values = static_cast<TValue *>(Memory::alloc_static(sizeof(TValue) * capacity));
-		hashes = static_cast<uint32_t *>(Memory::alloc_static(sizeof(uint32_t) * capacity));
-
-		for (uint32_t i = 0; i < capacity; i++) {
-			hashes[i] = EMPTY_HASH;
-		}
+	OAHashMap(uint32_t p_initial_capacity = 32) {
+		_resize_and_rehash(p_initial_capacity);
 	}
 
 	~OAHashMap() {
-		for (uint32_t i = 0; i < capacity; i++) {
-			if (hashes[i] == EMPTY_HASH) {
-				continue;
-			}
-
-			values[i].~TValue();
-			keys[i].~TKey();
-		}
-
-		Memory::free_static(keys);
-		Memory::free_static(values);
-		Memory::free_static(hashes);
+		_reset();
 	}
 };