Similar to #90082 and #90126.

**Benchmarks**
OAHashMap<int, int>
| Elements | Time insert before(usec) | Time insert after(usec) |
|----------|--------------------|-------------------|
| 5        | 4                  | 0                 |
| 25       | 0                  | 0                 |
| 125      | 8                  | 4                 |
| 625      | 42                 | 15                |
| 3125     | 261                | 116               |
| 15625    | 1781               | 565               |
| 78125    | 8986               | 2711              |
| 390625   | 41887              | 20942             |
| 1953125  | 303047             | 108786            |
| 9765625  | 1560636            | 637230            |


| Elements | Time find before(nsec) | Time find after(nsec) |
|----------|------------------|----------------|
| 10       | 304              | 368            |
| 50       | 727              | 647            |
| 250      | 3472             | 2605           |
| 1250     | 21905            | 15523          |
| 6250     | 143600           | 50483          |
| 31250    | 424157           | 300460         |
| 156250   | 2790448          | 1940480        |
| 781250   | 22462356         | 12027047       |
| 3906250  | 109030704        | 82341888       |
| 19531250 | 757996928        | 530154176      |