@Nazarwadim 2024-04-05
Are there any reasons why @Geometror didn't implement `fastmod` here? And can I add it here?