@Geometror 2024-04-05
Yeah, I think @AThousandShips is right, this is a legacy class. I think there might be some specific cases where it could be more performant than `HashMap`, but to be sure we should run some benchmarks (AStar and certain geometric operations use it). If I remember correctly we wanted to remove it at some point.