@Nazarwadim 2024-04-05
This data structure looks much better for trivial types than the standard one. We don't use extra 3 pointers that weigh 24 bytes in total, and we don't use extra system calls to allocate memory. It is also better friends with the cache.