@Nazarwadim 2024-04-05
Tested the regular and OA map from the master branch.
Time inserting before: 114961
Time inserting after: 141894 
 
Time inserting before: 111946
Time inserting after: 140684
 
Time inserting before: 110359
Time inserting after: 95678
 
Time find before: 54432136
Time find after: 51467400

The insertion is much worse because `MAX_OCCUPANCY` in hash_map is 0.75, and here it is 0.9.