@Nazarwadim 2024-04-06
Yes. I thought so because I did tests locally and they passed. This is most likely due to the refactoring commit.